Extract Checkout to another component
